+++ kernel/mm/string.c
@@ -82,7 +82,7 @@ int strcmp(const char *str1, const char *str2)
         }
         return (*str1 - *str2);
 }
-int strcmp_s(const char *str1, const char *str2, int n)
+int strcmp_s(const char *str1, const char *str2, size_t n)
 {
         while (n && *str1 && (*str1 == *str2)) {
                 str1++;
